+++ lib/papi-dts/papi.d.ts
@@ -549,17 +549,6 @@ declare module 'shared/utils/util' {
     },
     objId?: string,
   ): Set<string>;
-  /**
-   * Run an array of promises, and either return an array of the outcomes if they all were fulfilled
-   * or throw if at least one of them was rejected
-   *
-   * @param promises Array of promises to resolve
-   * @returns Array of `PromiseSettledResult` values from each promise if all promises were fulfilled.
-   *   Otherwise an exception will be thrown.
-   */
-  export function runPromisesAndThrowIfRejected(
-    ...promises: Promise<unknown>[]
-  ): Promise<PromiseSettledResult<unknown>[]>;
 }
 declare module 'shared/utils/papi-util' {
   import { ProcessType } from 'shared/global-this.model';
