+++ src/renderer/services/renderer-web-socket.service.ts
@@ -8,7 +8,7 @@ export const blockWebSocketsToPapiNetwork = (): void => {
   allowWebSocketsBackToPapiNetwork = false;
 };
 
-// There are lots of ways to reference localhost, and it's hard to catch them all
+// We are just filtering by port number on purpose to allow other connections to localhost
 function isPotentialConnectionToPapiNetwork(url: string | URL): boolean {
   let urlObj: URL;
   if (typeof url === 'string') {
@@ -17,16 +17,8 @@ function isPotentialConnectionToPapiNetwork(url: string | URL): boolean {
     urlObj = url;
   }
 
-  const { hostname, port } = urlObj;
-  // Check if the hostname is one of the common localhost values or the port number matches
-  return (
-    hostname === 'localhost' ||
-    hostname === '127.0.0.1' ||
-    hostname === '::1' ||
-    hostname.endsWith('.local') ||
-    hostname.endsWith('.localhost') ||
-    parseInt(port, 10) === parseInt(`${WEBSOCKET_PORT}`, 10)
-  );
+  const { port } = urlObj;
+  return parseInt(port, 10) === parseInt(`${WEBSOCKET_PORT}`, 10);
 }
 
 /**
